requirements.txt
asgiref==3.4.1
autopep8==1.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
dj-database-url==0.5.0
Django==3.2.11
gunicorn==20.1.0
idna==3.3
influxdb==5.3.1
msgpack==1.0.4
numpy==1.23.1
pandas==1.4.3
Pillow==9.2.0
pycodestyle==2.8.0
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
Unipath==1.1
urllib3==1.26.11
whitenoise==5.3.0
